+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
