+++ src/main/java/lotto/domain/RankCreator.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 import java.util.HashMap;
 import java.util.Map;
