+++ src/main/java/lotto/domain/WinningNumber.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
