+++ src/test/java/lotto/domain/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
