+++ src/test/java/lotto/domain/PayingMoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
