+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.domain;
 
 
 import org.assertj.core.api.Assertions;
